requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
datasets==3.2.0
dill==0.3.8
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pip==24.2
portalocker==2.8.2
propcache==0.4.1
psutil==5.9.8
pyarrow==15.0.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.12.0
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torchdata==0.7.1
torchvision==0.18.0
tqdm==4.66.4
transformers==4.44.2
typer-slim==0.19.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.24.0
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
onnxruntime>=1.17.0
onnx>=1.16.0

